requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
llama-index>=0.11.0
llama-index-llms-openai-like>=0.2.0
llama-index-embeddings-dashscope>=0.2.0
llama-index-readers-file>=0.2.0
langchain-community>=0.2.10
ragas>=0.1.14
pandas>=2.2.2
datasets>=2.20.0
pydantic>=2.7.0
